+++ cmd/magus-utils/main.go
@@ -22,6 +22,7 @@ var scribes = map[string]func(args []string) error{
 	"types":              runTypes,
 	"bindings":           runBindings,
 	"config":             runConfig,
+	"output":             runOutput,
 	"spells":             runSpells,
 	"sign":               runSign,
 	"api":                runAPI,
@@ -51,5 +52,5 @@ func main() {
 }
 
 func usage() {
-	fmt.Fprintln(os.Stderr, "usage: magus-utils <types|bindings|config|spells|sign|api|verify|cut|migrate|release-index|generate-changelog> [flags]")
+	fmt.Fprintln(os.Stderr, "usage: magus-utils <types|bindings|config|output|spells|sign|api|verify|cut|migrate|release-index|generate-changelog> [flags]")
 }